Authentication, Bugfix, Navbar and more
Added a AuthWrapper to protect certain routes as well as a way to get user information without needing to make a lot of requests to the backend.

I modified Layout as well to display a navbar if a user is provided (meaning a user has signed in).

Login and Logout now works fine, but might need some more improvements before deployment due to us not invalidating the old JWT, and just "removing" it...

Also created the default welcome page (it will render different things depending on user type, but currently it renders key information about the user).